Rework filters logic
