+++ src/components/custom-aggrid/custom-aggrid-filters/custom-aggrid-autocomplete-filter.tsx
@@ -4,21 +4,25 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-import React, { FunctionComponent, useCallback } from 'react';
+import React, { FunctionComponent, SyntheticEvent, useCallback } from 'react';
 import { Autocomplete, TextField } from '@mui/material';
 import { useLocalizedCountries } from '../../utils/localized-countries-hook';
 import { useIntl } from 'react-intl';
-import { useCustomAggridFilter } from './use-custom-aggrid-filter';
+import { useCustomAggridFilter } from '../hooks/use-custom-aggrid-filter';
 import { isStringOrNonEmptyArray } from '../custom-aggrid-header-utils';
 import { CustomAggridFilterProps } from './custom-aggrid-filter';
+import { FILTER_TEXT_COMPARATORS } from '../custom-aggrid-header.type';
 
 export const CustomAggridAutocompleteFilter: FunctionComponent<CustomAggridFilterProps> = ({ field, filterParams }) => {
     const { translate } = useLocalizedCountries();
     const intl = useIntl();
     const { filterOptions, isCountry, getEnumLabel } = filterParams;
 
-    const { selectedFilterData, autocompleteFilterParams } = useCustomAggridFilter(field, filterParams);
-    const { handleFilterAutoCompleteChange } = autocompleteFilterParams;
+    const { selectedFilterData, handleChangeFilterValue } = useCustomAggridFilter(field, filterParams);
+
+    const handleFilterAutoCompleteChange = (_: SyntheticEvent, data: string[]) => {
+        handleChangeFilterValue({ value: data, type: FILTER_TEXT_COMPARATORS.EQUALS });
+    };
 
     const getOptionLabel = useCallback(
         (option: string) =>
